+++ Sources/MacEverything/MacEverythingApp.swift
@@ -1,7 +1,16 @@
+import AppKit
 import SwiftUI
+
+final class AppDelegate: NSObject, NSApplicationDelegate {
+    func applicationDidFinishLaunching(_ notification: Notification) {
+        WindowActivation.activateApp()
+    }
+}
 
 @main
 struct MacEverythingApp: App {
+    @NSApplicationDelegateAdaptor(AppDelegate.self) private var appDelegate
+
     var body: some Scene {
         WindowGroup {
             ContentView()